+++ clients/derivatives-trading-portfolio-margin-pro/src/rest-api/types/index.ts
@@ -10,6 +10,7 @@ export * from './get-portfolio-margin-pro-account-balance-response-inner';
 export * from './get-portfolio-margin-pro-account-info-response';
 export * from './get-portfolio-margin-pro-span-account-info-response';
 export * from './get-portfolio-margin-pro-span-account-info-response-risk-unit-mmlist-inner';
+export * from './get-transferable-earn-asset-balance-for-portfolio-margin-response';
 export * from './mint-bfusd-for-portfolio-margin-response';
 export * from './portfolio-margin-collateral-rate-response';
 export * from './portfolio-margin-collateral-rate-response-inner';
@@ -26,3 +27,4 @@ export * from './query-portfolio-margin-pro-negative-balance-interest-history-re
 export * from './query-portfolio-margin-pro-negative-balance-interest-history-response-inner';
 export * from './redeem-bfusd-for-portfolio-margin-response';
 export * from './repay-futures-negative-balance-response';
+export * from './transfer-ldusdt-for-portfolio-margin-response';